be/Chatbot/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
sentence-transformers>=2.6.0
transformers>=4.36.0
huggingface-hub>=0.20.0
chromadb>=0.4.0
numpy>=1.26.0
torch>=2.0.0
